pdfwrite - permit Mod and CreDate pdfmarks in PDF 2.0
Bug 699807 "-dCompatibilityLevel=2.0` prevents all metadata from being set"

PDF 2.0 has deprecated The Document Info dictionary, so we prevented
pdfmarks from modifying the data in it, as a prelude to removing it
potentially when later versions of the PDF spec are approved.

However, as the report says, the Creation Date and Modification Date
are still required for documents with PieceInfo (only), so for now we
will permit those two values (only) to be set.