Bug 699810: Explicitly set stroke color for PDF Type 3 glyphs
When we render a cache-viable Type 3 glyphs (i.e. one that uses 'd1' rather
than 'd0'), ensure we set the stroke color to match the fill color.

This is necessary because if we do cache the glyph, we'll render to a mask, and
fill the mask. On the other hand, if we have to render the glyph uncached,
we have to ensure that the entire glyph is drawn with the same (fill) color, as
it would be when drawn cached.