Bug 699800: Fix scan converter problem seen with barcodes.
The new scan converter is incorrectly counting a pixel as being
touched if a line moves horizontally across the bottom edge of
that pixel.

This means that paths (for example) carefully constructed to cover
exactly a single pixel, such as:

   0 0 moveto 1 0 lineto 1 1 lineto 0 1 lineto closepath fill

(when rendered at 72dpi) will fill 2 pixels, rather than 1.

The fix is rather more involved that you'd like, as we need to
treat the cursor being on this vertical edge as a special case.
We flush and "null" the cursor when we arrive on such an edge,
and ignore horizontal motion across such edges.

The identical fixes are required both for trap and rect filling.