Bug 699857: don't push userdict in preparation for Type 1 fonts
Apparently, Fontographer fonts require a writable dictionary on the
dict stack. For some reason lost in mists of time, we pushed userdict for that
purpose - the problem is, that leaves any random redefinition of operators in
place when we interpret the font file.

Instead, we now push an empty, temporary dictionary for those Fontographer
fonts. As we already explicitly push systemdict (before the writable dictionary)
that ensures a consistent environment for the interpretation of Type 1 fonts.