PDF interpreter - limit page tree recusrsion checking
Bug #699856 "Attempting to open a carefully crafted PDF file results in long-running computation"

A sufficiently bad page tree can lead to us taking significant amounts
of time when checking the tree for recursion.

We can limit this by noting the number of pages in the root node
(given by /Count) and stopping the recursion check when we have
encountered that many leaf nodes.

Our other recursion checks work by reading the resources from the page
nodes and so are unaffected by this.