JPXDecode images kind of working (ha ha..)
Well, the images I have handy to test seem to work.

Parses the jpx image header, doesn't handle all the fields that are in the PS
version.  Prints warnings for missing things.

Hacks the colorspace mapping to something standard.

Wrote pdfi_JPX_filter (is it really this simple?)

Added jpx stuff to the makefile so it will compile correctly.

TODO: All the things that are missing.  Need more files to test.