Initial support for function arrays and function types 2 and 3
Type 2 (exponential interpolation) and type 3 (stitching) functions
as well as arrays of functions. There are bugs in the code, but at least
the initial implementation now runs.

Also, fix some minor errors in the make_float_array code which was still
behaving as a 'Domain' function, and a couple of memory cleanups in the
make_float_array and make_int_array functions.