PDF interpreter - don't use CIDFonts from the PDF file for missing ones
Bug #699937 "Poorly constructed PDF file can inappropriately reuse a CIDFont"

If we find a CIDFont is not defined (the FontDescriptor has no FontFile
entry) then we need to find a substitute. Normally this means going
to disk and either loading an explicit (user supplied) substitute, or
using a fallback CIDFont.

However, if the PDF file had defined a CIDFont with the same name
previously, we could use that instead. If the CIDFont was a complete
font this should be absolutely fine. However in this the CIDFont is a
*subset* font, but not identified as such. Using that font leads to
missing text....

This commit continues to use resourcestatus to locate a substitute
CIDFont, but it checks the 'status' flag. If its 0 then that means we
created it via an explicit defineresource, which means (I believe and
testing seems to confirm) that we created it from the PDF file. If its
1 or 2, then it was loaded from disk as a substitute (or is available
but not in VM, ie its on disk).

So if status is 0, don't use this CIDFont, it might not be reliable,
instead force an excursion to disk and load a substitute from there.

Tested with an explicit substitution defined in cidfmap to make sure
that this does not break at the same time. NB! If a PDF file does
define a CIDFont of the same name, then we won't (I believe) be able
to load the explicit CIDFont resource, we will always fall back to the
default.