Fix mswinpr2 device with subclassing
Bug # 699934 "mswinpr2 creates empty print jobs (with some printer drivers)"

The mswinpr2 device has a bizarre check in it, to see if the device is a
copy produced by copydevice. Its not clear to me why it does this, and
the code is very old. It looks to me like this should fail if the device
*is* a copy produced by copydevice, for the same reason it does with
subclassing, the Windows-specific parts are set to NULL.

I can't seem to provoke the device to go through this path in normal
operation, so I'm not going to try and remove it. Instead I've added
to the test a check to see if the device has been subclassed. If it has
we no longer reset the Windows-specific parts of the device, but carry
on using what we have.

While we are here, remove the annoying wdev macro which simply makes
it hard to debug. Especially when the macro has different definitions
in different parts of the source file!