Bug 699935: Change to allow GSView 5 to work with current Ghostscript.
In 2016, a change was made to consult the ProcessColorModel in order to
determine if PDF needed special Overprint Transparency handling, but for
some reason, the GSView display device doesn't have this paramemter in
the pagedevice dictionary. Change to rely on the /Colors parameter being
less than 4 for devices that don't need special Overprint trans handling.