PDF interpreter - bug in growPDFobjects when repairing PDF file
Bug #699970 "bug in PDF repair"

The function growPDFobjects is used to increase the size of the Objects,
ObjectStream, Generations and IsGlobal records when the number of
objects being recorded exceeds the size of the current record.

The problem was that the code assumes Generations is a string, which it
normally is. However when we are repairing a PDF file, we turn the
string into an array, and if we need to increase the size of the
record we need to make a larger array, not a larger string.

This commit checks the current type of the Generations record and either
creates a large string or larger array, working from the type of the
existing record.