Numerous fixes to function code.
Don't try to free Domain and Range if we didn't allcote them, but
copied them from the caller.

If we get the wrong number of entries in an Encod/Decode array then
return an error, not the number of entries we read.

If building a sub function fails, return an error, don't carry on.

As a bonus, the start of the limited parser for type 4 functions