Fix for color_info.comp_bits not getting setup correctly in ps2write device
Opening the device before the procs are filled in is not good.
This fix intentionally leaves some #ifdef'd out code, for future review.
The correct solution is probably more general.

(There is also a change this breaks something for other programs that use
plmain.c, but since this is on a branch for pdf, not too worried right now)

The problem being addressed is that check_device_separable() (called
from gs_opendevice()) needs to have the proc for "encode_color"
already setup.

The path through fs_setdevice_no_erase() does this correctly, as long
as the device has not been previously opened.

The actual bug I saw was in the ps2write device, which had
color_info.comp_bits set to 0 (because they were never initialized).