Bug 700023: Fix cups get/put_params LeadingEdge logic
Custom device LeadingEdge logic in complicated as it potentially requires
equivalent changes to changing the page size.

For a device like cups which merely wants the LeadingEdge value stored in its
own place it makes much more sense to allow the Ghostscript 'core' to handle
all the logic (which it has to do anyway to ensure the rendering is correct),
and then just copy the value into the cups specific place.

This saves the problem, as in this case, of unexpectedly ending up with two
LeadingEdge parameters being included in a call to put_params.