Allow cvrs to operate on 64-bit integers
Bug #700011 "cvrs may truncate integers"

zcvrs() cast the PostScript intval to an unsigned int internally, which
truncates 64-bit values to 32-bit values. Other similar operators are
capable of handling 64-bit values.

This commit casts the result to a ps_uint (PostScript unsigned int)
which varies in size depending on the platform. If a 64-bit integer
data type is available it will be 64-bit otherwise it will be 32-bit.

Additionally, check CPSI mode and if it is set, cast the ps_int to an
unsigned int in order to explicitly truncate to a 32-bit value. We use
CPSI compatibility mode in our internal testing, and if we don't do this
then two of the Quality Logic CET files throw errors.

Finally add some words to Use.htm regarding the word size of platforms
and some of the implications regarding 64-bit values in PostScript.