Fix pngalpha device with device subclassing
Bug #699815 "Segmentation fault for pdf input to pngalpha driver"

The pngalpha device, in pngalpha_create_buf_device() first creates an
in-memory rendering device (image32) and then overrides two of the
device methods with its own versions.

This doesn't work at all when we have subclassing devices loaded, because
the device that gets handed as the 'target' device is the head of the
chain of devices, and the pngalpha device simply pulls the methods
directly from that device. Note that this also wouldn't work if a
forwarding device were present.

We can work around it by descending the chain of devices, and pulling
the methods from the last device in the chain instead of using the
'target' directly. This still feels wrong but it should at least work.