Fixed xref parsing issue with Bug692177.pdf
The xref parser was assuming there would be a 0x0d or 0x0a at the end
of each entry, but as noted in Bug692177, this is not always the case.
In that file there is a "Prev" xref entry that has spaces where the line ends
should be.

The old code was decrementing the counter an unbounded amount until it
ran into an 0x0d, underflowing the buffer.  It caused random different
behaviors, none good (segfaults, thinking the file needed to be
repaired, etc).

After looking at the spec, it seems the format of these entries is
very specific, so just sticking a null after the required 18
characters should suffice.

I don't know if we need to be more robust to malformed entries, but
the old way wasn't really any more robust.  I suspect it might be
better to parse these entries more carefully, not using scanf, and
supporting any whitespace between each entry, but that's a total
rewrite (and probably can wait until we encounter a supposedly-valid
PDF with incorrectly formatted entries).