Change the way we read old style xref table entries.
If we don't find a line terminating character in the 20 bytes we read
from an old style xref table, attempt a much slower method of parsing it
by reading numbers.

Also, if sscanf fails to read 3 fields form the buffer, do the same slow
parsing.

Fixes bug 692177.pdf (and handles pclm.pdf) in a slower but hopefully
more robust fashion.

Ideally I'd prefer to throw an error and have the repair code rebuild
the xref, but the file is so badly constructed that the repair code
can't fix it.