Don't permit too many instances of the same filter
This is form the Sumatra 'rlebomb' file. The file is technically legal
but it includes multiple RLEDecode instances in the same Filter
array.

This leads to a very lengthy attempt to decode nonsense data. This
commit (which will likely need revision) limits the number of instances
of the same decode filter in a Filter array to 2.