Change graphics state initialisation for gs.
When we set the device for gs, trigger the graphics state to
initialise. This fix (credit to Chris) should hopefully solve
the differences we were seeing between gs and gpdl in the graphics
state setup.

Calling gpdl and gs with:

  -sDEVICE=ppmraw -r300 -dMaxBitmap=800000000

and a simple file that does a stroke was causing the stroking
code to be called with stroke_adjust = 1 in the gpdl case,
and stroke_adjust = 0 in the gs case.

After this fix, both are called with stroke_adjust = 0.

It is possible that this will resolve the differences in halftone
output we were seeing before.