Rejig gs_lib_ctx_t to allow sharing of core details.
gs_lib_ctx_t's now have a 'core' structure that contains things
like stdio and gs_id records.

When we try to make a gs_lib_ctx_t from a memory pointer that
doesn't contain a gs_lib_ctx_t, we do the same as now, and
create a completely new gs_lib_ctx, but put some of the contents
in a "core" structure within it.

When we try to make a gs_lib_ctx_t from a memory pointer that
DOES contain a gs_lib_ctx_t, we share the core structure with
the original.

This enables us to have 'families' of gs_lib_ctx_t's that share
the important things like stdin/stderr and gs_id records, while
keeping other things private (such as the io device tables which
are specific to individual languages).