Fix the b operators
The code test in b, b*, B and B* was incorrect, which meant that many
tests failed.

This caused problems because it meant that we created a gsave gstate
for every such operator, and never cleaned it up, leading to large
memory consumption.

In addition, we never reset the path, which meant that consecutive
operators led to a more and more complex path, again consuming huge
amounts of memory.

Finally, we didn't swap the colours from stroke to fill before running
the fill/eofill operations, which would have led to incorrect colours,
if we had actually ever got that far.

This (unsurprisingly) results in lots of diffs