Bug 700056: openjpeg allocator must return NULL if size too large.
Previously if an allocation larger than ARCH_MAX_UINT was requested
the size value was truncated and supplied to the gs allocator. Depending
on the values involved an allocation of this truncated value may
actually succeed. This allocation was then passed back to the openjpeg
library as claiming that the larger, initial allocation request was
satisfied, even though the allocation itself was actually smaller.
This resulted in segmentation faults.