GPDL: Cope with Postscript jobs that exit the server loop
In order to cope with jobs exit the server loop, we have to ensure that the
change from the nulldevice (used during gpdl initialization) and the 'real'
device being set prior to running the job, happens outside the server loop.

Otherwise the job's exiting of the server loop results in restoring to the
original state (with the nulldevice), and creating a new server loop context
from that base.

As things stand, the code will also exit the server loop to 'unset' the device
(i.e. revert back to the nulldevice) at the end of the job.

If we have to cope with a workflow that relies on making changes that really
persist between jobs, we'll have to rejig this to only unset the device
before shutting down.