PDF interpreter - Don't allow page stream fonts to be used for Annotations
Bug #700096 "Poorly constructed PDF file can inappropriately reuse a Font"

The file has a page stream which uses a Helvetica font which is embedded
and subset, but does not have a subset prefix on the name.

A later annotation uses Helvetica, but does not embed it. So we load
the Helvetica from the font cache. Unfortunately, this is the subset
font, and its Encoding is incompatible (as well as missing many glyphs)
so we get a lot of .notdefs.

Here we put a save and restore pair around the page contents, so that
when we come to the annotation rendering, we have restored the font
cache to an 'empty' state, so that we won't try to load any fonts that
were defined in the page stream to satisfy annotation requests.

For some peculiar reason one broken Sumatra file renders slightly
differently with this in place, and another was wrong before, differently
wrong now.