Avoid misleading warning message.
If a printer device provides its own dev_spec_op parameter, the
system prints a warning. This warning is intended to ensure that
devices remember to call down to the default printer one, but
appears irrespective of whether they do or not. This means it will
confuse most users.

Instead, introduce a debug only dso that we can use to check that
printer devices do indeed forward properly, and use that to print
the warning only when required.