Fixes for shading/functions crash, early interpreter exit
Found uninitialized stack variable that was causing the mischief
observed in Bug692216.pdf.  Need to make sure the params struct is set
to 0 in each of the functions where it appears on the stack
(especially because the structures are different sizes, so the
struct-copies don't completely initialize them).

Also found that the interpreter was exiting when the shading code
returned an error.  I added a check for ctx->pdfstoponerror in that
case, but I find that code pretty confusing as far as what the code is
supposed to be doing there.

These fixes make Bug692216.pdf work correctly, but introduce a new
crash on page 3 of Bug694909.pdf.  I have looked at this crash and I
think it is exposed because more code is running now, and not because
there is anything specifically wrong with the changes in this commit.
The new crash has something to do with pulling a value out of an image
dictionary that appears to be invalid (possibly already on the free
list -- it's type is not valid at the time it is fetched from the
dictionary) and then segfaulting when it tries to print out a debug
message complaining about this when it later frees it.