PDF interpreter - work around type 3 fonts with no FirstChar/LastChar
Bug #700130 "Invalid PDF file renders blank pages"

The file (and the customer original file) has a type 3 font with a
/Widths array, but neither of the required FirstChar or LastChar
entries.

With 9.25 and below this was leading to blank pages, where current code
continued. I think because the previous code left garbage on the stack.

This commit sets FirstChar to 0 and LastChar to the length of the
/Widths array if either is missing.