Update postscript initialisation to allow for -g and -r.
Using the the -g and -r flags, we need to set FIXEDMEDIA and
FIXEDRESOLUTION etc in the internals of the PS interpreter.

For this to work properly , it needs to be done during the init
phase (i.e. before gs_main_init2 is called).

Accordingly, we split the psapi init call in two. We also add
some plumbing to allow us to recover the resolutions/dimensions
of the pages from the top level, and to pass that down into
the internals of the interpreter.

For neatness, we make the standard gs init calls use these too.