Ensure stream and string persist as long as required
When we create an ICC based color space from a string source, we create a stream
object, with the stream data coming from that string.

Because the life of the ICC related objects created by the graphics library
is not dictated by the save/restore operations in the Postscript world, we
need to ensure that the stream and the string it references persist as long
as the graphics library needs.

To do this, we move the string data, and the stream object allocation into
"system VM", that is, garbage collected, but *not* save/restore memory.