Fix Bug 699806: Incorrect file size from "status" on Windows.
With commit afe8e546, I missed the use of _wstat and _stat in gp_mswin.c
for the "status" operator. Change to use _wstat64 and _stat64 and all
works. Tested with VS 2005 32-bit build and with VS 2015 32 and 64 bit
builds. Note the file size reported is correct for large files even on
a 32-bit build.