GPDL: Improve argument handling
We add a pl_interp_implementation function for setting a parameter
in the language. Only the PS implementation implements this currently,
and that sets the given parameters in systemdict.

While we parse arguments, we both write them into a parameter list
(which we feed to the device at the end of argument processing), and
we use the new call to pass them into the languages.

It would arguably be preferably for things to be written to the
device parameter list only if the device understood it, and the
rest could go to systemdict. This is impossible (or at least
very complicated) with the current implementation of param lists.
As it is, we end up doing pretty much the same as gs, which is
probably a good thing.

Also in this commit, we tidy up some of the arg handling; better
checks on radix numbers (both base and value), better arg
checking ("BATCHPLEASE" no longer matches for "BATCH" etc)