GPDL: Tweak gs_main_set_device
Do the actual device setting directly using C rather than entirely
in PS. This sidesteps the 'SAFER' checks and avoids problems in
the language switching where we try to insert a device under 'SAFER'
conditions.

This is not a security risk as we'll always be putting the same
device back in each time.