Fix infinite loop freeze running all PDF ATS files.
The infinite loop was caused by bumping the ref_count on a cache link
entry too early so that anytime we reached ICC_CACHE_MAXLINKS links in
the cache, we would loop forever.

This simplifies the code, removes a function that was only used once,
and also protects the clist multi-threading from potential lock up due
to more threads than slots in the icc_link_cache.

Add a semaphore when the link_cache is full (no zero ref_count entries)
to let other threads run and finish up with the links that are in the
cache.

No bug opened, but occurred with:
   gs -r600 -sDEVICE=ppmraw -o /dev/null PDF_1.7_ATS/*.pdf
NB: This requires skipping AC8ZD1CC_Save.pdf since that needs a password.