Rejig shadings and implement type 5-7
Pull most of the mesh shading code out of type 4 shadings, and make it
a generalised function for all mesh shadings to use.

Add type 5, 6 and 7 shadings.

This causes a regression in Bug692747.pdf when run through the pdfwrite
device, but multiple progressions, so for now I'm going to live with
the regression.