Bug 695813. Reduce memory usage for ICC profiles. Example AIX361DC_Save.pdf
The same ICC profiles and colorspace were used many times (in the example
the cited PDF_1.7_ATS file) which could result in memory usage growth and
performance impact associated with loading the same profile many times.
There was a icc_profile_cache in the gstate, but it was not being used for
ICCBased colorspaces, only for CIEBased* spaces.

Adding logic to cache ICCBased colorspaces and change ICCBased space to
be in stable memory so it can exist in the cache. This results in a
reduction of the RAM needed for the cited file from 1.0Gb down to 44Mb
and reduced the parse time from 29.4s down to 24.4s (overall time from
94.5s to 91.3s) for ppmraw 600 dpi.

Note that this optimization results in many fewer CS/cs operations being
emitted by pdfwrite, but this also can result in text being emitted as
TJ operations (glyphs with positioning) which can result in single pixel
shifts in glyphs. Sometimes this looks better, sometimes not, but the
reduction in size/complexity of the resulting PDF is worth it.