PDF interpreter - use Appearance streams for highlight annotations
We were previously never using the Appearance stream for a highlight
annotation, which is documented as incorrect in the PDF 2.0 spec.

The reason we were not using it is because a very few PDF files render
differently if we do. This is because Acrobat 'sometimes', and I cannot
figure out the conditions for htis, ignores a perfectly valid Appearance
and generates a new one.

In this case a customer produced a file where the QuadPoints entry is
incorrect. If we use the Appearance stream (as Acrobat does) then the
rendering is correct. If we generate an appearance then, because the
QuadPoints are incorrect, the rendering is incorrect.

Since the PDF 2.0 specification now makes it clear that we should be
using a supplied Appearance stream, I've chosen to change the behaviour
so that we do.

I've also made the text for an appearance stream slightly more robust
and general.

This does produce a few diffs, because of the afore-mentioned files. In
my opinion Acrobat is incorrect in how it treats these files. Our new
rendering is correct.