PS interpreter - check infinite result on exp, where possible
Bug #700181 "using exp with 0 as base" (again)

The reporter for the bug failed to mention this condition in the
original report. Technically we don't always have isinf() available
(I suspect all real compilers do these days) but where we do we can
check the result. Where we don't, we're stuck. This mimics the
behaviour of other math operations in Ghostscript.