Coverity ID 323322
The new 'transform_pixel_region' method default for forwarding devices
tests the target device to see if its NULL (some forwarding devices are
not always forwarding devices, eg bbox), and if it is NULL then instead
of passing the call to the target device it calls the graphics library
default.

Unfortunately it was passing the target device as an argument, and this
is now certain to be NULL. In gx_default_transform_pixel_region_begin()
we then dereference the pointer which would cause a crash.

So instead, pass the device itself when its not actually a forwarding
device.

Also, some white space changes.