ps2write - move the page level save/restore wrapper
Bug #700232 "A5 duplex to duplex-capable Postscript printer prints single-sided"

Commit 0f6067d2531298060392d0e25fa759d320e03021 for bug #697245
removed a duplicate definition of pagesave, because it transpires that
Adobe interpreters don't free save objects when they are no longer
referenced. Previously we had defined /pagesave in both the page
content stream, and the code for altering page size.

The problem is that we alter media size as part of processing the
'stream' keyword, which is after we have defined /pagesave. Previously
we would then have redefined it if we needed to alter the media size.

When we got to the end of the page (after the showpage) we would then
restore the save object. Because we had defined the save object after we
had altered the media size the media would remain unchanged. With the
commit above, however, the media size would be restored back to the
original.

When we process page 2, the media size would then need to he altered
again, requiring a setpagedevice to be executed. The action of
setpagedevice, when Duplex is true, is to eject any unpaired pages and
start again with the first side.

Thus Duplex was working entirely correctly, simply not as desired.

This commit moves the pagesave and restore into the 'stream' and
'endstream' processing, so that is taken care of after any media size
changes.